REFACTOR test classes to `Feature` test namespace
